refactor: add unique component folders to the Main folder
